Finished basic functions

+++ panel/Panel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Formats.Asn1;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -76,20 +77,57 @@ namespace test_liste.panel
         // Removes cars.
         public void RemoveCars()
         {
-            Console.WriteLine("Type what tag to remove cars by :\n");
-            string tag = Console.ReadLine();
-            _carService.RemoveByTag(tag);
-            Console.WriteLine("\nRemoved cars succesfully!\nDo you want to display the new car list? (Y/N) :\n");
-            string choice = Console.ReadLine();
-            Console.WriteLine();
-            switch (choice)
+            Console.WriteLine("What do you want to remove cars by :\n");
+            Console.WriteLine("Id - Remove a car that has a specific id.");
+            Console.WriteLine("Tag - Remove all cars with a certain tag.\n");
+
+            string choice1 = Console.ReadLine();
+
+            int count = 0;
+
+            switch (choice1.ToLower())
             {
-                case "Y":
-                    _carService.Display();
+                case "id":
+                    Console.WriteLine("\nType the id of the car you want removed :\n");
+                    int id = Console.Read();
+                    if (_carService.FindById(id) != null)
+                    {
+                        _carService.RemoveById(id);
+                        count = 1;
+                    }                   
+                    break;
+                case "tag":
+                    Console.WriteLine("\nType what tag to remove cars by :\n");
+                    string tag = Console.ReadLine();
+                    count = _carService.SearchForTag(tag).Count();
+                    if(count > 0)
+                    {
+                        _carService.RemoveByTag(tag);
+                    }                    
                     break;
                 default:
                     break;
             }
+
+            if (count > 0)
+            {
+                Console.WriteLine($"\nRemoved {count} cars succesfully!\nDo you want to display the new car list? (Y/N) :\n");
+                string choice2 = Console.ReadLine();
+                Console.WriteLine();
+                switch (choice2.ToLower())
+                {
+                    case "y":
+                        _carService.Display();
+                        break;
+                    default:
+                        break;
+                }
+            }
+            else
+            {
+                Console.WriteLine("\nNo cars found by specified parameters.\n");
+            }
+            
         }
     }
 }
